+++ CHANGES.rst
@@ -14,8 +14,8 @@ Development
 -----
 
 * Breaking changes: ``cache_factory`` keyword arguments ``depends_on`` is a
-  list of callables which receiving request as only argument. Callables return
-  the dependency to be identified by cache manager.
+  list of callables which receiving request as positional argument. Callables
+  return the dependency to be identified by cache manager.
 * Add arguments ``predicates`` to ``cache_factory`` which permits to add
   predicates to cache key used for the view (useful for query strings).